requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0

# Database
sqlalchemy>=2.0.40
pymysql==1.1.0
cryptography==42.0.0

# File handling
python-multipart==0.0.6
openpyxl==3.1.2
aiofiles==23.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scheduler
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
hypothesis==6.96.1

# Email
aiosmtplib==3.0.1
